requirements.txt
# Core dependencies for AI Job Chommie
# Minimal set for quick startup

# Web frameworks
fastapi==0.104.1
uvicorn[standard]==0.25.0
aiohttp==3.9.1

# Database and caching
redis==5.0.1
sqlalchemy==2.0.23

# ML essentials
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
numpy==1.26.2
pandas==2.1.4

# Utils
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2
psutil==5.9.6
pyyaml==6.0.1

# Monitoring
prometheus-client==0.19.0

# Security
cryptography==41.0.7
